tests: add planet scene inference testing
select a gpu_id via parameters. In case of the test argument not being set,
a random ID between 0 and 7 is selected. As this only uses an environment
variable to set the gpu_id and not any cuda/torch facilities, it should be
painless on cpu-only systems in terms of configuration.